requirements.txt
# requirements.txt

Django==4.2.0
Pillow==10.4.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
google-generativeai==0.3.1